Move icon into mainthreadwebview
